+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add Ordering to Filters
 - [MSSQL ONLY] Add ability to perform Regex redactions on data loads and existing catalogues
+- Add RAW Table Date Column Override for Delta Loads
 - Fix Delta Load off by one issue
 - Update Migration strategy to account for all Primary Keys when moving from staging -> live
 - Fix UI issue with viewing cross-database SQL results